Feature - move SQL dependency tracking to sep project
