+++ src/Arcus.Observability.Telemetry.Core/Extensions/ILoggerExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
 using System.Globalization;
 using System.Net;
 using System.Net.Http;
@@ -595,45 +594,6 @@ namespace Microsoft.Extensions.Logging
             bool isSuccessful = (int) statusCode >= 200 && (int) statusCode < 300;
 
             logger.LogInformation(HttpDependencyFormat, targetName, dependencyName, (int) statusCode, duration, startTime.ToString(CultureInfo.InvariantCulture), isSuccessful, context);
-        }
-
-        /// <summary>	
-        ///     Logs a SQL dependency	
-        /// </summary>	
-        /// <param name="logger">Logger to use</param>	
-        /// <param name="connectionString">SQL connection string</param>	
-        /// <param name="tableName">Name of table</param>	
-        /// <param name="operationName">Name of the operation that was performed</param>	
-        /// <param name="isSuccessful">Indication whether or not the operation was successful</param>	
-        /// <param name="measurement">Measuring the latency to call the SQL dependency</param>	
-        /// <param name="context">Context that provides more insights on the dependency that was measured</param>
-        [Obsolete("Will be moved to separate package 'Arcus.Observability.Telemetry.Sql' in the future")]
-        public static void LogSqlDependency(this ILogger logger, string connectionString, string tableName, string operationName, DependencyMeasurement measurement, bool isSuccessful, Dictionary<string, object> context = null)
-        {
-            Guard.NotNull(logger, nameof(logger));
-            Guard.NotNull(measurement, nameof(measurement));
-
-            LogSqlDependency(logger, connectionString, tableName, operationName, measurement.StartTime, measurement.Elapsed, isSuccessful, context);
-        }
-
-        /// <summary>	
-        ///     Logs a SQL dependency	
-        /// </summary>	
-        /// <param name="logger">Logger to use</param>	
-        /// <param name="connectionString">SQL connection string</param>	
-        /// <param name="tableName">Name of table</param>	
-        /// <param name="operationName">Name of the operation that was performed</param>	
-        /// <param name="startTime">Point in time when the interaction with the HTTP dependency was started</param>	
-        /// <param name="duration">Duration of the operation</param>	
-        /// <param name="isSuccessful">Indication whether or not the operation was successful</param>	
-        /// <param name="context">Context that provides more insights on the dependency that was measured</param>
-        [Obsolete("Will be moved to separate package 'Arcus.Observability.Telemetry.Sql' in the future")]
-        public static void LogSqlDependency(this ILogger logger, string connectionString, string tableName, string operationName, DateTimeOffset startTime, TimeSpan duration, bool isSuccessful, Dictionary<string, object> context = null)
-        {
-            Guard.NotNullOrEmpty(connectionString, nameof(connectionString));
-            var connection = new SqlConnectionStringBuilder(connectionString);
-
-            LogSqlDependency(logger, connection.DataSource, connection.InitialCatalog, tableName, operationName, isSuccessful, startTime, duration, context);
         }
 
         /// <summary>
